// app/types.ts
export enum View {
    ABOUT_US = 'ABOUT_US',
    OUR_TEAM = 'OUR_TEAM',
    FAQS = 'FAQS',
    CONTACT_SUPPORT = 'CONTACT_SUPPORT',
    SIGNUP = 'SIGNUP',
    DASHBOARD = 'DASHBOARD',
    PRODUCT_LIST = 'PRODUCT_LIST',
    INQUIRY_LIST = 'INQUIRY_LIST',
    BUSINESS_INFO = 'BUSINESS_INFO',
    MESSAGES = 'MESSAGES',
    NOTIFICATIONS = 'NOTIFICATIONS',
    SETTINGS = 'SETTINGS',
    POLICIES = 'POLICIES',
    HOME = 'HOME',
    MY_ORDERS = 'MY_ORDERS',
    WISHLIST = 'WISHLIST',
    MY_QUOTES = 'MY_QUOTES',
    CATEGORIES = 'CATEGORIES',
    CHAT_SESSION = 'CHAT_SESSION',
    POST_REQUIREMENT = 'POST_REQUIREMENT'
}